feat: Refactor moderation profile page
